Struct maxminddb::Reader

pub struct Reader<S: AsRef<[u8]>> {
    pub metadata: Metadata,
    // some fields omitted
}

A reader for the MaxMind DB format. The lifetime 'data is tied to the lifetime of the underlying buffer holding the contents of the database file.

Fields
metadata: Metadata
Implementations

impl<'de> Reader<Vec<u8>>

pub fn open_readfile<P: AsRef<Path>>(
database: P
) -> Result<Reader<Vec<u8>>, MaxMindDBError>

Open a MaxMind DB database file by loading it into memory.

Example

let reader = maxminddb::Reader::open_readfile("test-data/test-data/GeoIP2-City-Test.mmdb").unwrap();

impl<'de, S: AsRef<[u8]>> Reader<S>

pub fn from_source(buf: S) -> Result<Reader<S>, MaxMindDBError>

Open a MaxMind DB database from anything that implements AsRef<u8>

Example

use std::fs;
let buf = fs::read("test-data/test-data/GeoIP2-City-Test.mmdb").unwrap();
let reader = maxminddb::Reader::from_source(buf).unwrap();

pub fn lookup<T>(&'de self, address: IpAddr) -> Result<T, MaxMindDBError> where
T: Deserialize<'de>,

Lookup the socket address in the opened MaxMind DB

Example:

use maxminddb::geoip2;
use std::net::IpAddr;
use std::str::FromStr;

let reader = maxminddb::Reader::open_readfile("test-data/test-data/GeoIP2-City-Test.mmdb").unwrap();

let ip: IpAddr = FromStr::from_str("89.160.20.128").unwrap();
let city: geoip2::City = reader.lookup(ip).unwrap();
print!("{:?}", city);